[REFACTOR] 리뷰 반영하여 비즈니스 코드 리팩토링

+++ src/main/java/com/sports/server/command/league/domain/LeagueTeamRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface LeagueTeamRepository extends JpaRepository<LeagueTeam, Integer> {
+public interface LeagueTeamRepository extends JpaRepository<LeagueTeam, Long> {
     Optional<LeagueTeam> findByLeagueAndTeam(League league, Team team);
 
     @Query("SELECT lt.team.id FROM LeagueTeam lt WHERE lt.league.id = :leagueId AND lt.team.id IN :teamIds")
